requirements.txt
asgiref==3.7.2
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
Django==5.0.1
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.14.0
docutils==0.19
filelock==3.13.1
idna==3.4
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.4
mysql-connector-python==8.1.0
mysqlclient==2.2.4
Pillow==9.5.0
platformdirs==4.1.0
protobuf==4.21.12
psycopg2==2.9.9
Pygments==2.15.0
pypiwin32==223
pytz==2023.3.post1
pywin32==306
requests==2.28.2
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.25.0
WooCommerce==3.0.0
